Move docker.mk to a separate folder, so it can be re-used by other languages later
