+++ internal/artifacts/golang/projectfiles.go
@@ -14,7 +14,6 @@ func ProjectFiles() []*artifacts.ProjectFile {
 		artifacts.NewProjectFile(gotemplates.VersionTxtFile, "VERSION.txt", goTemplateDelimiters),
 		artifacts.NewProjectFile(gotemplates.Dockerfile, "Dockerfile", goTemplateDelimiters),
 		artifacts.NewProjectFile(gotemplates.VersionGo, "version/version.go", goTemplateDelimiters),
-		artifacts.NewProjectFile(gotemplates.DockerMkFile, "docker.mk", goTemplateDelimiters),
 		artifacts.NewProjectFile(gotemplates.MainGo, "main.go", goTemplateDelimiters),
 		artifacts.NewProjectFile(gotemplates.GitIgnore, ".gitignore", goTemplateDelimiters),
 	}
